adding configuration elements for s3
