+++ lib/db_blaster.rb
@@ -9,6 +9,7 @@ require 'db_blaster/configuration'
 require 'db_blaster/source_table_configuration'
 require 'db_blaster/source_table_configuration_builder'
 require 'db_blaster/publisher'
+require 'db_blaster/sns_publisher'
 require 'db_blaster/publish_source_table'
 require 'db_blaster/chunker'
 require 'db_blaster/finder_sql'
