+++ lib/db_blaster/publisher.rb
@@ -2,7 +2,7 @@
 
 # publish records to SNS topic
 module DbBlaster
-  # Publishes records to AWS SNS
+  # Base class for publishing
   class Publisher
     attr_reader :source_table, :records
 
@@ -12,31 +12,11 @@ module DbBlaster
     end
 
     def self.publish(source_table, records)
-      new(source_table, records).publish
+      SnsPublisher.new(source_table, records).publish
     end
 
     def publish
-      topic.publish(message_attributes: message_attributes,
-                    message: records.to_json)
-    end
-
-    def topic
-      resource.topic(DbBlaster.configuration.sns_topic)
-    end
-
-    def resource
-      @resource ||= Aws::SNS::Resource.new(client: client)
-    end
-
-    def client
-      @client ||= Aws::SNS::Client.new(region: DbBlaster.configuration.aws_region,
-                                       credentials: Aws::Credentials.new(DbBlaster.configuration.aws_access_key,
-                                                                         DbBlaster.configuration.aws_access_secret))
-    end
-
-    def message_attributes
-      (DbBlaster.configuration.extra_sns_message_attributes || {})
-        .merge('source_table' => { data_type: 'String', string_value: source_table.name })
+      raise NotImplementedError
     end
   end
 end
